requirements.txt
python-telegram-bot>=20.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
pandas>=2.0.0
numpy>=1.24.0
aiofiles>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
